FIX: 기존 API 응답의 코스 길이, 고도값을 자연수 형태로 수정 (#103)
코스 조회와 관련된 API 응답에 포함된 코스 길이와 고도값을 자연수 형태로 수정했습니다.

+++ src/main/java/com/server/running_handai/domain/course/dto/CourseDetailDto.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record CourseDetailDto(
         Long courseId,
         String courseName,
-        double distance,
+        int distance,
         int duration,
         int minElevation,
         int maxElevation,
@@ -25,7 +25,7 @@ public record CourseDetailDto(
         return new CourseDetailDto(
                 course.getId(),
                 course.getName(),
-                course.getDistance(),
+                (int) course.getDistance(),
                 course.getDuration(),
                 (int) course.getMinElevation().doubleValue(),
                 (int) course.getMaxElevation().doubleValue(),
